created function to add a new department

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE employees (
   id INTEGER NOT NULL,
   first_name VARCHAR(30),
   last_name VARCHAR (30),
-  role_id INTEGER (200),
+  role_id INTEGER (10),
   manager_id INTEGER (10) NULL,
   PRIMARY KEY (id)
 );
